Autopopulate JWT with jQuery =( (=
This script assumes you have a POST route named login. It listens for a response and as long as "error" isn't in the response it saves the text content of the response field in localstorage. It assumes that the only thing returned from the login route is the JWT, no JSON object.

This doesn't check which site you're using so if you load up a new API, the authentication fields will be autopopulated with a JWT from another project.